+++ backend/document/create_confirm_doc.go
@@ -15,6 +15,6 @@ func createConfirmDoc(writer http.ResponseWriter, request *http.Request) {
 			return
 		}
 		doConfirm(id, tx, writer)
-		con.SendAccept(uint(id), writer)
+		con.SendAccept(id, writer)
 	}
 }
